requirements.txt
flask
sqlalchemy
flask_sqlalchemy
pymysql
psycopg2-binary
pyyaml
pytz
pytest
mysqlclient==2.1.1